chore: remove tsOverride on ae runs
