+++ api-extractor.json
@@ -77,7 +77,7 @@
 		 * SUPPORTED TOKENS: <projectFolder>, <packageName>, <unscopedPackageName>
 		 * DEFAULT VALUE: "<projectFolder>/tsconfig.json"
 		 */
-		// "tsconfigFilePath": "<projectFolder>/tsconfig.json",
+		"tsconfigFilePath": "<projectFolder>/tsconfig.docs.json"
 		/**
 		 * Provides a compiler configuration that will be used instead of reading the tsconfig.json file from disk.
 		 * The object must conform to the TypeScript tsconfig schema:
@@ -88,52 +88,7 @@
 		 *
 		 * DEFAULT VALUE: no overrideTsconfig section
 		 */
-		"overrideTsconfig": {
-			"compilerOptions": {
-				// Type Checking
-				"allowUnreachableCode": false,
-				"allowUnusedLabels": false,
-				"exactOptionalPropertyTypes": true,
-				"noFallthroughCasesInSwitch": true,
-				"noImplicitOverride": true,
-				"noImplicitReturns": true,
-				"noPropertyAccessFromIndexSignature": false,
-				"noUncheckedIndexedAccess": true,
-				"noUnusedLocals": true,
-				"noUnusedParameters": true,
-				"strict": true,
-
-				// Modules
-				"allowArbitraryExtensions": false,
-				"allowImportingTsExtensions": false,
-				"module": "ESNext",
-				"moduleResolution": "Bundler",
-				"resolveJsonModule": true,
-				"resolvePackageJsonExports": false,
-				"resolvePackageJsonImports": false,
-
-				// Emit
-				"declaration": true,
-				"declarationMap": true,
-				"importHelpers": false,
-				"newLine": "lf",
-				"noEmitHelpers": true,
-				"outDir": "dist",
-				"rootDir": "./src",
-				"removeComments": false,
-				"sourceMap": true,
-
-				// Interop Constraints
-				"forceConsistentCasingInFileNames": true,
-				"isolatedModules": true,
-
-				// Language and Environment
-				"experimentalDecorators": true,
-				"lib": ["ESNext"],
-				"target": "ES2022",
-				"useDefineForClassFields": true
-			}
-		}
+		// "overrideTsconfig": {}
 		/**
 		 * This option causes the compiler to be invoked with the --skipLibCheck option. This option is not recommended
 		 * and may cause API Extractor to produce incomplete or incorrect declarations, but it may be required when
