+++ packages/formatters/api-extractor.json
@@ -2,12 +2,5 @@
 	"extends": "../../api-extractor.json",
 	"docModel": {
 		"projectFolderUrl": "https://github.com/discordjs/discord.js/tree/main/packages/formatters"
-	},
-	"compiler": {
-		"overrideTsconfig": {
-			"compilerOptions": {
-				"types": ["node"]
-			}
-		}
 	}
 }
