+++ packages/structures/api-extractor.json
@@ -2,13 +2,5 @@
 	"extends": "../../api-extractor.json",
 	"docModel": {
 		"projectFolderUrl": "https://github.com/discordjs/discord.js/tree/main/packages/structures"
-	},
-	"compiler": {
-		"overrideTsconfig": {
-			"compilerOptions": {
-				"types": ["node"]
-			},
-			"exclude": ["src/**/*.d.ts"]
-		}
 	}
 }
