+++ packages/structures/package.json
@@ -5,7 +5,7 @@
 	"description": "Wrapper around Discord's structures",
 	"scripts": {
 		"build": "tsc && inject-version",
-		"build:docs": "tsc -p tsconfig.docs.json && cpy \"./src/*.d.ts\" \"./dist-docs\"",
+		"build:docs": "tsc -p tsconfig.docs.json",
 		"test": "vitest run --config ../../vitest.config.ts",
 		"lint": "prettier --check . && cross-env TIMING=1 eslint --format=pretty src",
 		"format": "prettier --write . && cross-env TIMING=1 eslint --fix --format=pretty src",
