+++ packages/structures/src/channels/DMChannel.ts
@@ -1,6 +1,5 @@
 import type { APIDMChannel, ChannelType } from 'discord-api-types/v10';
-import { Mixin } from '../Mixin.js';
-import type { MixinTypes } from '../MixinTypes.d.ts';
+import { Mixin, type MixinTypes } from '../Mixin.js';
 import type { Partialize } from '../utils/types.js';
 import { Channel } from './Channel.js';
 import { ChannelPinMixin } from './mixins/ChannelPinMixin.js';
