+++ packages/structures/src/index.ts
@@ -21,4 +21,4 @@ export * from './subscriptions/index.js';
 export * from './Mixin.js';
 export * from './utils/optimization.js';
 export type * from './utils/types.js';
-export type * from './MixinTypes.d.ts';
+export type * from './Mixin.js';
